embed-service/requirements.txt
fastapi
uvicorn
sentence-transformers
pillow
python-multipart